Fix for bug 63830: Only delete edited Draft mail if "com.openexchange.mail.deleteDraftOnTransport" is set to "true"

+++ com.openexchange.mail.compose.impl/src/com/openexchange/mail/compose/impl/CompositionSpaceServiceImpl.java
@@ -577,7 +577,7 @@ public class CompositionSpaceServiceImpl implements CompositionSpaceService {
                     }
                 }
                 MailPath editFor = meta.getEditFor();
-                if (null != editFor) {
+                if (null != editFor && MailProperties.getInstance().isDeleteDraftOnTransport(session.getUserId(), session.getContextId())) {
                     try {
                         mailInterface.deleteMessages(editFor.getFolderArgument(), new String[] { editFor.getMailID() }, true);
                     } catch (Exception e) {
